Factor out graph converter
