+++ crates/paralegal-flow/src/utils/mod.rs
@@ -24,13 +24,13 @@ use crate::{
         ty,
     },
     rustc_span::ErrorGuaranteed,
-    Either, HashMap, HashSet, Symbol, TyCtxt,
+    Either, HashSet, Symbol, TyCtxt,
 };
 
 pub use flowistry_pdg_construction::FnResolution;
 
 use std::cmp::Ordering;
-use std::{cell::RefCell, default::Default, hash::Hash, pin::Pin};
+use std::hash::Hash;
 
 pub mod resolve;
 
@@ -1040,47 +1040,6 @@ pub fn with_temporary_logging_level<R, F: FnOnce() -> R>(filter: log::LevelFilte
     r
 }
 
-/// This code is adapted from [`flowistry::cached::Cache`] but with a recursion
-/// breaking mechanism. This alters the [`Self::get`] method signature to return
-/// an [`Option`] of a reference. In particular the method will return [`None`]
-/// if it is called *with the same key* while computing a construction function
-/// for that key.
-pub struct RecursionBreakingCache<In, Out>(RefCell<HashMap<In, Option<Pin<Box<Out>>>>>);
-
-impl<In, Out> RecursionBreakingCache<In, Out>
-where
-    In: Hash + Eq + Clone,
-    Out: Unpin,
-{
-    pub fn size(&self) -> usize {
-        self.0.borrow().len()
-    }
-    /// Get or compute the value for this key. Returns `None` if called recursively.
-    pub fn get<'a>(&'a self, key: In, compute: impl FnOnce(In) -> Out) -> Option<&'a Out> {
-        if !self.0.borrow().contains_key(&key) {
-            self.0.borrow_mut().insert(key.clone(), None);
-            let out = Pin::new(Box::new(compute(key.clone())));
-            self.0.borrow_mut().insert(key.clone(), Some(out));
-        }
-
-        let cache = self.0.borrow();
-        // Important here to first `unwrap` the `Option` created by `get`, then
-        // propagate the potential option stored in the map.
-        let entry = cache.get(&key).unwrap().as_ref()?;
-
-        // SAFETY: because the entry is pinned, it cannot move and this pointer will
-        // only be invalidated if Cache is dropped. The returned reference has a lifetime
-        // equal to Cache, so Cache cannot be dropped before this reference goes out of scope.
-        Some(unsafe { std::mem::transmute::<&'_ Out, &'a Out>(&**entry) })
-    }
-}
-
-impl<In, Out> Default for RecursionBreakingCache<In, Out> {
-    fn default() -> Self {
-        Self(RefCell::new(HashMap::default()))
-    }
-}
-
 pub fn time<R, F: FnOnce() -> R>(msg: &str, f: F) -> R {
     info!("Starting {msg}");
     let time = std::time::Instant::now();
@@ -1151,5 +1110,16 @@ impl<'tcx> Spanned<'tcx> for (LocalDefId, mir::Location) {
     fn span(&self, tcx: TyCtxt<'tcx>) -> Span {
         let body = tcx.body_for_def_id(self.0).unwrap();
         (&body.body, self.1).span(tcx)
+    }
+}
+
+pub fn map_either<A, B, C, D>(
+    either: Either<A, B>,
+    f: impl FnOnce(A) -> C,
+    g: impl FnOnce(B) -> D,
+) -> Either<C, D> {
+    match either {
+        Either::Left(l) => Either::Left(f(l)),
+        Either::Right(r) => Either::Right(g(r)),
     }
 }
